% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{url_parse}
\alias{url_parse}
\title{Parses a URL string into its components.}
\usage{
url_parse(url)
}
\arguments{
\item{url}{The URL string to parse.}
}
\value{
A list containing the components of the URL: scheme, user, password, host, path, raw_path, query, raw_query, and fragment.
}
\description{
Parses a URL string into its components.
}
\examples{
library(urlparse)
url_parse("https://host.com/path?query#fragment")

}
